Create serial_processor
Updated interface to serial connector. One thing to note is that the Nano 33 BLE Sense is somewhat sensitive so if there are issues, hard restart the board and you should be good to go. Check the serial monitor output just to be sure but make sure to close it to not hang up your COM port.